painel-esus/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
APScheduler==3.10.4
astroid==3.0.2
async-timeout==4.0.3
attrs==23.2.0
black==24.4.2
blinker==1.6.2
blueprint-decr==0.1.1
cachelib==0.9.0
Cerberus==1.3.5
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
coverage==7.3.2
cryptography==41.0.4
customtkinter==5.2.2
darkdetect==0.8.0
dill==0.3.7
distlib==0.3.7
engineering-notation==0.10.0
exceptiongroup==1.1.3
filelock==3.12.4
Flask==2.3.3
Flask-APScheduler==1.13.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
frozenlist==1.4.1
greenlet==2.0.2
identify==2.5.29
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.24.4
packaging==23.1
pandas==2.0.3
pathspec==0.12.1
platformdirs==3.10.0
pluggy==1.3.0
polars==1.5.0
pre-commit==3.4.0
psycopg2-binary==2.9.7
pyarrow==17.0.0
pycparser==2.21
PyJWT==2.8.0
pylint==3.0.3
pytest==7.4.2
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.21
tk==0.1.0
tk-tools==0.16.0
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.5
virtualenv==20.24.5
Werkzeug==2.3.7
yarl==1.9.4
zipp==3.17.0
